move counter to fastmetrics
