+++ libs/tendermint/p2p/metrics.go
@@ -4,6 +4,7 @@ import (
 	"github.com/go-kit/kit/metrics"
 	"github.com/go-kit/kit/metrics/discard"
 	"github.com/go-kit/kit/metrics/prometheus"
+	"github.com/okex/exchain/libs/tendermint/libs/fastmetrics"
 	stdprometheus "github.com/prometheus/client_golang/prometheus"
 )
 
@@ -47,13 +48,13 @@ func PrometheusMetrics(namespace string, labelsAndValues ...string) *Metrics {
 			Name:      "peers",
 			Help:      "Number of peers.",
 		}, labels).With(labelsAndValues...),
-		PeerReceiveBytesTotal: NewCounterFrom(stdprometheus.CounterOpts{
+		PeerReceiveBytesTotal: fastmetrics.NewCounterFrom(stdprometheus.CounterOpts{
 			Namespace: namespace,
 			Subsystem: MetricsSubsystem,
 			Name:      "peer_receive_bytes_total",
 			Help:      "Number of bytes received from a given peer.",
 		}, append(labels, "peer_id", "chID")).With(labelsAndValues...),
-		PeerSendBytesTotal: NewCounterFrom(stdprometheus.CounterOpts{
+		PeerSendBytesTotal: fastmetrics.NewCounterFrom(stdprometheus.CounterOpts{
 			Namespace: namespace,
 			Subsystem: MetricsSubsystem,
 			Name:      "peer_send_bytes_total",
